+++ Rubicon.Core/Chart/TargetSwitch.cs
@@ -10,7 +10,7 @@ namespace Rubicon.Core.Chart;
     /// <summary>
     /// The time to trigger this switch. Stored on disk in measures.
     /// </summary>
-    [Export] public double Time = 0.0;
+    [Export] public float Time = 0f;
     
     /// <summary>
     /// The bar line's name to switch to.
@@ -20,7 +20,7 @@ namespace Rubicon.Core.Chart;
     /// <summary>
     /// The time converted from measures to milliseconds. Should be ignored when serialized.
     /// </summary>
-    public double MsTime = 0d;
+    public float MsTime = 0f;
     
     /// <summary>
     /// Converts the pre-existing Time and Length variables to milliseconds and stores them in MsTime and MsLength, using the provided BpmInfo.
